Block Inspection

inspect_block creates a list where the original data has been divided into blocks denoted in the state vector. Streamlines the process of inspecting specific blocks of interest.

inspect_block(data, block_length)

Arguments

data
data

block_length
integer value to divide data

Value

A list where each item is a data frame that contains the original data for each block denoted in the state vector.

See also

Examples

inspect_block(security_logs, 30)

#> [[1]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 McAfee        NSP            Attempt       223.7… 32.73…      135    20000
#>  2 CISCO         ASA            Failure       174.1… 219.1…       80       25
#>  3 IBM           SNIPS          Success       174.1… 145.1…       80      593
#>  4 McAfee        NSP            Success       227.1… 151.1…       21       25
#>  5 Juniper       SRX            Success       28.9.… 145.1…      135       90
#>  6 McAfee        NSP            Success       28.9.… 145.1…       80       25
#>  7 McAfee        NSP            Attempt       28.9.… 56.13…       80       25
#>  8 McAfee        ePO            Attempt       223.7… 151.1…       25    20000
#>  9 McAfee        ePO            Attempt       174.1… 151.1…       25      593
#> 10 CISCO         ASA            Attempt       227.1… 56.13…       25       90
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[2]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 Juniper       SRX            Failure       227.1… 151.1…       80       90
#>  2 CISCO         ASA            Success       28.9.… 32.73…       25    20000
#>  3 McAfee        ePO            Success       174.1… 32.73…      113       80
#>  4 CISCO         ASA            Failure       227.1… 219.1…       21       80
#>  5 Juniper       SRX            Success       227.1… 219.1…       25      593
#>  6 CISCO         ASA            Failure       28.9.… 219.1…      113       90
#>  7 CISCO         ASA            Failure       89.13… 145.1…       21    20000
#>  8 Palo Alto Ne… Firewall       Failure       174.1… 145.1…      113       90
#>  9 IBM           SNIPS          Success       174.1… 151.1…       21      593
#> 10 CISCO         ASA            Success       89.13… 145.1…       80       80
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[3]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 IBM           SNIPS          Success       28.9.… 219.1…      113       25
#>  2 Palo Alto Ne… Firewall       Success       89.13… 56.13…       21       80
#>  3 Palo Alto Ne… Firewall       Success       89.13… 145.1…       21       25
#>  4 IBM           SNIPS          Failure       174.1… 151.1…       80       80
#>  5 CISCO         ASA            Attempt       223.7… 219.1…      113       90
#>  6 CISCO         ASA            Success       223.7… 219.1…      113       25
#>  7 Juniper       SRX            Attempt       89.13… 151.1…      113       80
#>  8 McAfee        NSP            Attempt       223.7… 219.1…       80      593
#>  9 McAfee        NSP            Failure       89.13… 219.1…       21    20000
#> 10 McAfee        NSP            Failure       89.13… 151.1…       21       25
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[4]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 Juniper       SRX            Success       223.7… 219.1…       80       80
#>  2 CISCO         ASA            Success       223.7… 32.73…       21       80
#>  3 McAfee        ePO            Attempt       227.1… 56.13…      135       80
#>  4 IBM           SNIPS          Success       28.9.… 32.73…      135      593
#>  5 Juniper       SRX            Attempt       223.7… 56.13…       25       90
#>  6 McAfee        ePO            Success       174.1… 151.1…      135       25
#>  7 IBM           SNIPS          Attempt       89.13… 32.73…       21      593
#>  8 McAfee        NSP            Success       227.1… 151.1…       80       90
#>  9 CISCO         ASA            Success       223.7… 56.13…       80      593
#> 10 McAfee        ePO            Success       28.9.… 151.1…       80      593
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[5]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 Palo Alto Ne… Firewall       Success       89.13… 32.73…      135       25
#>  2 McAfee        NSP            Failure       223.7… 32.73…       80    20000
#>  3 McAfee        NSP            Attempt       89.13… 56.13…      135    20000
#>  4 Palo Alto Ne… Firewall       Success       89.13… 151.1…       25       25
#>  5 IBM           SNIPS          Attempt       174.1… 145.1…      135    20000
#>  6 IBM           SNIPS          Success       174.1… 32.73…       25       80
#>  7 McAfee        NSP            Failure       174.1… 151.1…       25       90
#>  8 McAfee        NSP            Success       227.1… 219.1…      113       80
#>  9 IBM           SNIPS          Success       227.1… 219.1…       21       80
#> 10 Palo Alto Ne… Firewall       Attempt       174.1… 145.1…       80    20000
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[6]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 CISCO         ASA            Success       223.7… 145.1…       80       25
#>  2 McAfee        ePO            Attempt       174.1… 219.1…       21    20000
#>  3 IBM           SNIPS          Attempt       223.7… 151.1…       25      593
#>  4 CISCO         ASA            Success       223.7… 151.1…      113       90
#>  5 IBM           SNIPS          Attempt       174.1… 151.1…       21       25
#>  6 IBM           SNIPS          Attempt       223.7… 145.1…      113       90
#>  7 McAfee        NSP            Attempt       223.7… 151.1…       21       90
#>  8 Juniper       SRX            Success       227.1… 219.1…       80       90
#>  9 CISCO         ASA            Attempt       28.9.… 32.73…      113       90
#> 10 CISCO         ASA            Success       28.9.… 56.13…      113       80
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[7]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 Palo Alto Ne… Firewall       Success       227.1… 32.73…      135       80
#>  2 CISCO         ASA            Success       227.1… 219.1…      135      593
#>  3 McAfee        ePO            Success       89.13… 145.1…       25    20000
#>  4 CISCO         ASA            Attempt       227.1… 145.1…       25       25
#>  5 Juniper       SRX            Failure       174.1… 151.1…       25       80
#>  6 Palo Alto Ne… Firewall       Success       227.1… 56.13…       25       80
#>  7 Palo Alto Ne… Firewall       Attempt       223.7… 56.13…       21       25
#>  8 IBM           SNIPS          Attempt       89.13… 219.1…       25       25
#>  9 McAfee        ePO            Attempt       89.13… 56.13…       25    20000
#> 10 McAfee        NSP            Success       28.9.… 32.73…      135       90
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[8]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 IBM           SNIPS          Attempt       227.1… 56.13…       25       80
#>  2 CISCO         ASA            Failure       227.1… 219.1…       21       80
#>  3 McAfee        NSP            Success       89.13… 56.13…       80       80
#>  4 CISCO         ASA            Attempt       174.1… 151.1…       21       90
#>  5 McAfee        ePO            Success       28.9.… 219.1…      113    20000
#>  6 CISCO         ASA            Failure       28.9.… 56.13…      135      593
#>  7 McAfee        NSP            Attempt       174.1… 145.1…      113       25
#>  8 Juniper       SRX            Failure       227.1… 151.1…       21       25
#>  9 McAfee        NSP            Failure       223.7… 32.73…       21       25
#> 10 Juniper       SRX            Failure       227.1… 56.13…       80      593
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[9]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 IBM           SNIPS          Attempt       227.1… 56.13…       21    20000
#>  2 Palo Alto Ne… Firewall       Attempt       89.13… 56.13…       25    20000
#>  3 McAfee        ePO            Success       223.7… 219.1…       80       25
#>  4 Juniper       SRX            Failure       227.1… 56.13…       21       25
#>  5 Juniper       SRX            Success       174.1… 145.1…       21       80
#>  6 Palo Alto Ne… Firewall       Failure       28.9.… 145.1…       80       80
#>  7 McAfee        NSP            Success       28.9.… 151.1…      113    20000
#>  8 McAfee        NSP            Failure       227.1… 32.73…       25       80
#>  9 Palo Alto Ne… Firewall       Success       227.1… 151.1…       21       80
#> 10 CISCO         ASA            Success       28.9.… 145.1…      113       90
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#> 
#> [[10]]
#> # A tibble: 30 x 10
#>    Device_Vendor Device_Product Device_Action Src_IP Dst_IP Src_Port Dst_Port
#>    <chr>         <chr>          <chr>         <chr>  <chr>     <int>    <int>
#>  1 McAfee        NSP            Failure       227.1… 145.1…       25       80
#>  2 Juniper       SRX            Attempt       174.1… 145.1…       21      593
#>  3 McAfee        ePO            Success       28.9.… 56.13…      113      593
#>  4 IBM           SNIPS          Success       227.1… 32.73…       25       90
#>  5 Juniper       SRX            Failure       223.7… 151.1…       80       80
#>  6 Juniper       SRX            Failure       174.1… 56.13…       21       90
#>  7 IBM           SNIPS          Success       223.7… 56.13…       80       90
#>  8 McAfee        NSP            Success       223.7… 219.1…       21       80
#>  9 McAfee        ePO            Failure       28.9.… 219.1…       25       80
#> 10 McAfee        ePO            Attempt       174.1… 56.13…      135       25
#> # … with 20 more rows, and 3 more variables: Protocol <chr>, Country_Src <chr>,
#> #   Bytes_TRF <int>
#>