Kaiser's Index of Factorial Simplicity

kaisers_index computes scores designed to assess the quality of a factor analysis solution. It measures the tendency towards unifactoriality for both a given row and the entire matrix as a whole. Kaiser proposed the evaluations of the score shown below:

In the .90s: Marvelous

In the .80s: Meritorious

In the .70s: Middling

In the .60s: Mediocre

In the .50s: Miserable

< .50: Unacceptable

Use as basis for selecting original or rotated loadings/scores in factor_analysis.

kaisers_index(loadings)

Arguments

loadings
numerical matrix of the factor loadings

Value

Vector containing the computed score

References

H. F. Kaiser, "An index of factorial simplicity," Psychometrika, vol. 39, no. 1, pp. 31-36, 1974.

See also

Examples

# Perform Factor Analysis with matrix \code{x}
x <- matrix(rnorm(200*3), ncol = 10)

x %>%
  horns_curve() %>%
  factor_analysis(x, hc_points = .) %>%
  factor_analysis_results(fa_loadings_rotated) %>%
  kaisers_index()

#> [1] 0.8162322