Principal Component Analysis

principal_components relates the data to a set of a components through the eigen-decomposition of the correlation matrix, where each component explains some variance of the data and returns the results as an object of class prcomp.

principal_components(data, retx = TRUE, center = TRUE, scale. = FALSE,
  tol = NULL, ...)

Arguments

data
numeric data.

retx
a logical value indicating whether the rotated variables should be returned.

center
a logical value indicating whether the variables should be shifted to be zero centered. Alternately, a vector of length equal the number of columns of x can be supplied. The value is passed to scale.

scale.
a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place. The default is FALSE for consistency with S, but in general scaling is advisable. Alternatively, a vector of length equal the number of columns of data can be supplied. The value is passed to scale.

tol
a value indicating the magnitude below which components should be omitted. (Components are omitted if their standard deviations are less than or equal to tol times the standard deviation of the first component.) With the default null setting, no components are omitted. Other settings for tol could be tol = 0 or tol = sqrt(.Machine$double.eps), which would omit essentially constant components.

...
arguments passed to or from other methods.

Value

principal_components returns a list containing the following components:

pca_sdev: the standard deviations of the principal components (i.e., the square roots of the eigenvalues of the correlation matrix, though the calculation is actually done with the singular values of the data matrix).

pca_loadings: the matrix of variable loadings (i.e., a matrix whose columns contain the eigenvectors).

pca_rotated: if retx is TRUE the value of the rotated data (the centred (and scaled if requested) data multiplied by the rotation matrix) is returned. Hence, cov(x) is the diagonal matrix diag(sdev^2).

pca_center: the centering used

pca_scale: whether scaling was used

Details

The calculation is done by a singular value decomposition of the (centered and possibly scaled) data matrix, not by using eigen on the covariance matrix. This is generally the preferred method for numerical accuracy

See also

Examples

x <- matrix(rnorm(200 * 3), ncol = 10)
principal_components(x)

#> $pca_sdev
#>  [1] 1.2962713 1.2558586 1.2132392 1.1542074 1.0238156 0.9643643 0.8933342
#>  [8] 0.8811908 0.7982203 0.6015554
#> 
#> $pca_loadings
#>                PC1         PC2         PC3         PC4          PC5         PC6
#>  [1,]  0.381246317 -0.14691494  0.44409047 -0.01472215  0.090793971 -0.36385905
#>  [2,]  0.620548857  0.28641005 -0.52854451  0.29681554  0.008660355  0.04205106
#>  [3,]  0.189064241  0.37244007  0.39156170 -0.22235348 -0.123353942 -0.09486165
#>  [4,]  0.115588639 -0.02951389  0.06545116  0.18525384  0.906698011  0.08710650
#>  [5,] -0.137209770 -0.13882505 -0.22189600  0.22406329 -0.020727078  0.14340477
#>  [6,] -0.127173689 -0.60760795 -0.09922383  0.32154560 -0.137211064 -0.09402970
#>  [7,]  0.398897837 -0.30625142  0.43984707  0.40848512 -0.212563384  0.26527003
#>  [8,]  0.005809644  0.32702829  0.11142481  0.24560778 -0.219859758  0.54246491
#>  [9,]  0.387241376 -0.15519536 -0.30582249 -0.18983415 -0.192287021 -0.35826614
#> [10,] -0.276325088  0.37884002  0.08560608  0.64117667 -0.067909382 -0.57171192
#>               PC7         PC8         PC9        PC10
#>  [1,] -0.16245439  0.35660432 -0.27556749 -0.51526856
#>  [2,]  0.20112936  0.28766285  0.18517523 -0.07806113
#>  [3,]  0.48938892 -0.43694540  0.34652128 -0.21924129
#>  [4,]  0.08993283 -0.30494905 -0.11124202  0.06340277
#>  [5,] -0.39471280 -0.35105047  0.34344843 -0.66668801
#>  [6,]  0.66015095 -0.04415552 -0.11644496 -0.14975184
#>  [7,] -0.22129070 -0.11946889  0.26861520  0.36853379
#>  [8,]  0.06335024 -0.14185270 -0.64455116 -0.19517159
#>  [9,] -0.18417851 -0.58246329 -0.37009141  0.14461099
#> [10,] -0.08404826 -0.08557807 -0.03516651  0.12082283
#> 
#> $pca_rotated
#>               PC1          PC2        PC3         PC4         PC5         PC6
#>  [1,] -1.51830616  0.556601356 -0.6147185  1.92054759 -1.56385907  0.59871359
#>  [2,]  0.94365533  0.009358914  1.2125091 -0.82819432  0.36851848 -0.30122514
#>  [3,]  0.25003939 -0.047105962 -1.0350509 -0.39209025 -1.00606517 -1.83022329
#>  [4,]  0.21231179  0.312162465 -1.0839665  1.68108439 -0.37390079  0.38740856
#>  [5,]  0.07405330  1.998852154 -0.6614454 -1.73665782  0.04443971  0.56889959
#>  [6,]  0.11572635  0.460548127 -0.3085498 -1.48622086  1.46379867  1.02054632
#>  [7,] -1.19967905 -0.050166906 -0.8146942  1.29545291  1.56153593 -1.09397591
#>  [8,] -0.96197620  0.796695635 -1.2395702  0.34287383  0.25067637  0.53622186
#>  [9,]  0.49397175  0.783572639 -0.6811486 -0.96401642  0.52774678 -0.28779352
#> [10,]  0.56745539 -2.019350237 -1.0890808 -0.58079698 -1.19242858  0.98531018
#> [11,]  2.65341928  1.182516975  0.8013031  1.12831201 -0.78412145  0.03213318
#> [12,]  0.42920752  1.201776273 -1.6436036 -0.21342994  0.95912347 -1.06760627
#> [13,]  1.43427512 -1.739939088  1.7250906  1.98990652 -0.96249470  1.28384165
#> [14,]  0.95694038  1.186829607 -1.3038756  0.72561841  0.15333550  0.42621718
#> [15,]  1.23549509  0.317435477  0.5122638 -0.35293091 -0.71723143  0.72615276
#> [16,]  0.30909780  0.088278237  2.1650970 -0.93703788  0.67664526  0.48878506
#> [17,] -0.41707455  1.053207144 -0.4558909 -1.59814019 -0.63163445  0.83112810
#> [18,]  0.38840922  2.219865620 -0.4103199 -0.88917497  0.85832784 -0.92362045
#> [19,] -0.72401195 -0.444510304 -0.7412773 -0.17350248 -0.64349336  2.19914139
#> [20,] -1.40521406 -1.784678778 -3.2912227 -0.88030405  0.54732545 -1.16998750
#> [21,]  1.87451359 -0.375325392 -0.8954098  0.04728083  1.46601952 -0.15321064
#> [22,] -0.05501274  1.007068972 -1.4394039 -0.02515316  2.38330343 -1.41573736
#> [23,] -1.30754940  0.973708812 -0.2724064 -1.02177421 -0.98137717 -1.41394170
#> [24,]  1.64248395 -1.747364850 -1.4470448  0.59425432  0.64931148 -0.92394215
#> [25,]  0.85816084 -0.596249864  0.3834663  1.55947781 -0.83210022 -0.98486963
#> [26,] -0.89325885  1.119253640 -0.3450632  1.55919830  0.10260396  0.31563814
#> [27,]  0.86164294  0.760571544  0.8484896 -0.48337949 -0.05836657 -0.69044093
#> [28,] -1.57037492 -2.101418765 -1.9632240 -1.03386401 -1.84409908 -0.77556309
#> [29,] -0.22412763 -1.411368746  0.2560071 -0.74981639 -0.09833102  0.67453766
#> [30,] -1.30088255 -1.813952301 -0.4995960  1.14307118  0.49711992  0.34038673
#> [31,] -1.90115094  1.978029537  1.8105212 -1.39684974  1.04893046  0.30751870
#> [32,] -1.99606113 -1.586548736  0.4252285  0.10700421  0.57748194  0.42595160
#> [33,]  0.34550733  1.884762769  0.1906796 -0.93304971 -1.10060947  0.82200797
#> [34,]  1.69209822 -1.323073871  1.9500316 -0.04577172 -0.59258486 -1.30977154
#> [35,] -1.59792753 -1.634747060 -0.8612015  0.51149483  0.52311689  1.54174540
#> [36,]  1.38154962 -1.244709937  0.4325525 -2.29513578 -0.64725745  0.79800605
#> [37,] -0.32633226 -2.283312279  0.9569437  0.05038291  1.14225380 -2.30443785
#> [38,]  1.68946459  1.713597997 -0.6169675 -0.23887191 -0.37805479  0.03984386
#> [39,]  0.69684199 -0.587290779 -1.0806599 -0.04893910 -1.81322604  0.02863340
#> [40,]  0.50431211 -2.408296181  1.1266273 -1.68601512  1.63554309 -1.11269373
#> [41,]  1.33972255 -1.080841357  0.9567162 -0.04040834 -0.08757263  0.65068265
#> [42,] -1.94570462  0.669407607  0.9232580  0.63069483 -0.65346579 -0.13597110
#> [43,]  1.44064527  0.880437028 -0.5687700  1.23467055 -0.39706674 -0.88948534
#> [44,]  2.90212101  1.750250788 -1.0305339  1.34728294  0.03967435 -0.03530228
#> [45,] -0.91465346  1.098857696  0.4902522  0.11231686  0.05381678  0.23640247
#> [46,]  0.18862750  1.205692066  0.8748675  1.95428319 -0.04751048 -0.61419631
#> [47,] -0.35886467 -0.606824893  0.1892159  1.27444523 -0.76407567 -0.86410225
#> [48,] -0.35654384 -0.705832402 -1.3367786  1.76330989 -0.47787725  0.83398340
#> [49,]  0.86849112  0.270897413  1.5549070  1.66160903  0.78507908  0.38687798
#> [50,] -1.53459518 -0.630037898  2.2925030  2.61898184  1.26419160  0.09823967
#> [51,] -3.41626883  0.603845749  0.6286157 -0.64591993  0.62815707 -0.13630060
#> [52,]  1.33921123 -0.645621319 -0.3529361 -1.14041519  0.93141610 -0.32091502
#> [53,] -2.37952719  0.908613247  1.0620676  0.76765074 -0.22093819 -0.30568717
#> [54,] -1.38198152  1.061000220 -0.5827538 -0.60326340 -2.85129055 -0.92831880
#> [55,] -0.02112469 -0.765740331  1.0213768  0.48006122  0.68121177  0.42094173
#> [56,] -0.14017729  0.403028320  2.5433807 -1.68785846 -1.88256940 -1.36242109
#> [57,] -1.49609868  0.417133595  1.7872431 -0.78494746 -0.45417647  0.37033958
#> [58,]  0.10298247  0.022670091 -1.9887771 -0.49580721  0.92223187  2.37757536
#> [59,]  1.13895881 -2.228722390  0.7411863 -1.57117365 -0.52254147  1.01890428
#> [60,]  0.41308705  0.966502912  0.7935402 -0.54035530  1.83738372  1.57902462
#>               PC7          PC8          PC9         PC10
#>  [1,]  0.98920906 -1.331247680  0.508643904 -0.558712412
#>  [2,] -0.85576561 -1.129282797  0.144336880 -0.768719879
#>  [3,] -0.92551597 -0.373169048  0.186912325  1.097958380
#>  [4,] -1.40512163  1.256106203  0.123238416 -0.350130607
#>  [5,] -0.49146272 -1.344626735 -0.541325297  0.452166422
#>  [6,]  0.72068968 -0.926956564  0.857095527  1.302577771
#>  [7,]  1.43428010  0.479939666  0.973674916 -0.043150399
#>  [8,]  0.90370657  0.798994402 -0.298070142 -0.209618373
#>  [9,] -0.81731662  0.799031333 -0.657299190 -0.254716551
#> [10,] -0.59032850 -0.089061087 -1.105492741 -0.998304398
#> [11,]  0.34511587 -1.287230786 -0.670604158 -0.228742596
#> [12,]  0.07331302  0.139958468 -1.664122606 -0.395238834
#> [13,]  0.39250461 -0.514100049  0.752311126 -0.163175896
#> [14,]  0.24849963  0.311767068  0.969243311 -0.798920908
#> [15,]  1.82835634 -0.124406574 -0.288850652  1.240149949
#> [16,]  0.91412225  0.734282572 -0.008845660  1.182477563
#> [17,] -0.69243206 -0.785266633 -0.356551158 -0.139739180
#> [18,]  1.36635932 -0.365949043  0.840084718  0.589267603
#> [19,]  2.19444004 -0.066299476 -0.451920021 -0.474378407
#> [20,] -0.37835670  1.087099578 -1.308957126  0.256211611
#> [21,]  0.12778261 -1.342554871 -0.849878848  0.330479304
#> [22,]  1.62611592 -0.064568606  0.032445586 -0.645426881
#> [23,] -0.06286973 -0.019728139  0.866785472  0.328579989
#> [24,] -0.96794785  1.390022018  0.293868802 -0.027166406
#> [25,]  0.23894069  0.649384379 -0.872425219  1.008489631
#> [26,] -0.82494359  0.252357125 -0.096754265  0.393249366
#> [27,] -1.16517555  1.749731756  0.138662584 -0.132709829
#> [28,] -0.22586621  0.046764454  0.489808919  0.293159414
#> [29,]  0.87435787  0.716903874 -1.138288982 -0.779501752
#> [30,]  0.89021883 -0.910750827 -0.192355016 -0.203391872
#> [31,] -1.82342798 -0.358241439  0.117821311 -0.001922600
#> [32,] -1.23932477 -2.099763624 -1.242235863 -0.054210012
#> [33,] -1.12139172  2.272347216 -0.155082399 -0.224653294
#> [34,]  0.50583047 -0.239282416  0.492136275 -0.299121892
#> [35,] -1.21780551  0.317335353  0.885792839  1.176967370
#> [36,]  0.23734362  0.136867847 -0.179574396  0.051067803
#> [37,]  0.01639921 -1.405271295  0.005555418 -0.460807981
#> [38,] -0.20249013  0.316473534  0.421572407 -0.381565413
#> [39,] -1.18134303 -1.499746165  0.293351467  1.204602485
#> [40,] -0.22582890  0.182050821  1.889823263 -0.598917736
#> [41,]  0.48130957  0.110606259 -0.401420536 -0.273412410
#> [42,] -0.61589785 -0.611265142  0.379344511 -1.705436185
#> [43,] -0.06432790 -0.005455767  1.584044212 -0.662372354
#> [44,] -0.34898831 -1.107076642 -1.158139239  0.051831807
#> [45,]  1.06804119  0.535589515 -0.005886766  0.346033397
#> [46,] -0.22173727  0.372618645  0.591879902  0.632369191
#> [47,] -0.09313797  0.371046253 -2.057585722  0.233755510
#> [48,] -0.53368590  0.154201217  1.414596275  0.525846445
#> [49,] -0.88012229  0.497272187 -0.012382038  0.139239429
#> [50,]  0.08393584  1.171553486 -1.084392251  0.349883273
#> [51,]  0.57140487  0.437874072 -0.317405141  0.239117305
#> [52,] -0.03010354 -0.443525844  0.228758730 -0.292823145
#> [53,] -0.99623699 -1.765952513  0.641996474 -0.001615218
#> [54,]  1.28129289  0.559939493  0.793314943 -0.595600189
#> [55,]  0.82339140  0.775359245 -0.012639176  0.220031782
#> [56,]  1.01178796  0.111808522 -0.960244171  0.089287760
#> [57,] -0.21173219  0.634829907 -0.364366320 -0.311778138
#> [58,]  0.56317308 -0.211843540  0.506915950 -0.363862621
#> [59,] -0.28074920  0.854726448  1.077541599  0.086772933
#> [60,] -1.12048832  0.197780386 -0.048462965 -0.421729125
#> 
#> $pca_center
#>  [1] -0.19528802  0.35736637  0.20852324  0.02817919  0.08022248 -0.07598368
#>  [7]  0.01316175  0.08751549 -0.05739960 -0.05836311
#> 
#> $pca_scale
#> [1] FALSE
#> 

principal_components(x, scale = TRUE)

#> $pca_sdev
#>  [1] 1.2962713 1.2558586 1.2132392 1.1542074 1.0238156 0.9643643 0.8933342
#>  [8] 0.8811908 0.7982203 0.6015554
#> 
#> $pca_loadings
#>                PC1         PC2         PC3         PC4          PC5         PC6
#>  [1,]  0.381246317 -0.14691494  0.44409047 -0.01472215  0.090793971 -0.36385905
#>  [2,]  0.620548857  0.28641005 -0.52854451  0.29681554  0.008660355  0.04205106
#>  [3,]  0.189064241  0.37244007  0.39156170 -0.22235348 -0.123353942 -0.09486165
#>  [4,]  0.115588639 -0.02951389  0.06545116  0.18525384  0.906698011  0.08710650
#>  [5,] -0.137209770 -0.13882505 -0.22189600  0.22406329 -0.020727078  0.14340477
#>  [6,] -0.127173689 -0.60760795 -0.09922383  0.32154560 -0.137211064 -0.09402970
#>  [7,]  0.398897837 -0.30625142  0.43984707  0.40848512 -0.212563384  0.26527003
#>  [8,]  0.005809644  0.32702829  0.11142481  0.24560778 -0.219859758  0.54246491
#>  [9,]  0.387241376 -0.15519536 -0.30582249 -0.18983415 -0.192287021 -0.35826614
#> [10,] -0.276325088  0.37884002  0.08560608  0.64117667 -0.067909382 -0.57171192
#>               PC7         PC8         PC9        PC10
#>  [1,] -0.16245439  0.35660432 -0.27556749 -0.51526856
#>  [2,]  0.20112936  0.28766285  0.18517523 -0.07806113
#>  [3,]  0.48938892 -0.43694540  0.34652128 -0.21924129
#>  [4,]  0.08993283 -0.30494905 -0.11124202  0.06340277
#>  [5,] -0.39471280 -0.35105047  0.34344843 -0.66668801
#>  [6,]  0.66015095 -0.04415552 -0.11644496 -0.14975184
#>  [7,] -0.22129070 -0.11946889  0.26861520  0.36853379
#>  [8,]  0.06335024 -0.14185270 -0.64455116 -0.19517159
#>  [9,] -0.18417851 -0.58246329 -0.37009141  0.14461099
#> [10,] -0.08404826 -0.08557807 -0.03516651  0.12082283
#> 
#> $pca_rotated
#>               PC1          PC2        PC3         PC4         PC5         PC6
#>  [1,] -1.51830616  0.556601356 -0.6147185  1.92054759 -1.56385907  0.59871359
#>  [2,]  0.94365533  0.009358914  1.2125091 -0.82819432  0.36851848 -0.30122514
#>  [3,]  0.25003939 -0.047105962 -1.0350509 -0.39209025 -1.00606517 -1.83022329
#>  [4,]  0.21231179  0.312162465 -1.0839665  1.68108439 -0.37390079  0.38740856
#>  [5,]  0.07405330  1.998852154 -0.6614454 -1.73665782  0.04443971  0.56889959
#>  [6,]  0.11572635  0.460548127 -0.3085498 -1.48622086  1.46379867  1.02054632
#>  [7,] -1.19967905 -0.050166906 -0.8146942  1.29545291  1.56153593 -1.09397591
#>  [8,] -0.96197620  0.796695635 -1.2395702  0.34287383  0.25067637  0.53622186
#>  [9,]  0.49397175  0.783572639 -0.6811486 -0.96401642  0.52774678 -0.28779352
#> [10,]  0.56745539 -2.019350237 -1.0890808 -0.58079698 -1.19242858  0.98531018
#> [11,]  2.65341928  1.182516975  0.8013031  1.12831201 -0.78412145  0.03213318
#> [12,]  0.42920752  1.201776273 -1.6436036 -0.21342994  0.95912347 -1.06760627
#> [13,]  1.43427512 -1.739939088  1.7250906  1.98990652 -0.96249470  1.28384165
#> [14,]  0.95694038  1.186829607 -1.3038756  0.72561841  0.15333550  0.42621718
#> [15,]  1.23549509  0.317435477  0.5122638 -0.35293091 -0.71723143  0.72615276
#> [16,]  0.30909780  0.088278237  2.1650970 -0.93703788  0.67664526  0.48878506
#> [17,] -0.41707455  1.053207144 -0.4558909 -1.59814019 -0.63163445  0.83112810
#> [18,]  0.38840922  2.219865620 -0.4103199 -0.88917497  0.85832784 -0.92362045
#> [19,] -0.72401195 -0.444510304 -0.7412773 -0.17350248 -0.64349336  2.19914139
#> [20,] -1.40521406 -1.784678778 -3.2912227 -0.88030405  0.54732545 -1.16998750
#> [21,]  1.87451359 -0.375325392 -0.8954098  0.04728083  1.46601952 -0.15321064
#> [22,] -0.05501274  1.007068972 -1.4394039 -0.02515316  2.38330343 -1.41573736
#> [23,] -1.30754940  0.973708812 -0.2724064 -1.02177421 -0.98137717 -1.41394170
#> [24,]  1.64248395 -1.747364850 -1.4470448  0.59425432  0.64931148 -0.92394215
#> [25,]  0.85816084 -0.596249864  0.3834663  1.55947781 -0.83210022 -0.98486963
#> [26,] -0.89325885  1.119253640 -0.3450632  1.55919830  0.10260396  0.31563814
#> [27,]  0.86164294  0.760571544  0.8484896 -0.48337949 -0.05836657 -0.69044093
#> [28,] -1.57037492 -2.101418765 -1.9632240 -1.03386401 -1.84409908 -0.77556309
#> [29,] -0.22412763 -1.411368746  0.2560071 -0.74981639 -0.09833102  0.67453766
#> [30,] -1.30088255 -1.813952301 -0.4995960  1.14307118  0.49711992  0.34038673
#> [31,] -1.90115094  1.978029537  1.8105212 -1.39684974  1.04893046  0.30751870
#> [32,] -1.99606113 -1.586548736  0.4252285  0.10700421  0.57748194  0.42595160
#> [33,]  0.34550733  1.884762769  0.1906796 -0.93304971 -1.10060947  0.82200797
#> [34,]  1.69209822 -1.323073871  1.9500316 -0.04577172 -0.59258486 -1.30977154
#> [35,] -1.59792753 -1.634747060 -0.8612015  0.51149483  0.52311689  1.54174540
#> [36,]  1.38154962 -1.244709937  0.4325525 -2.29513578 -0.64725745  0.79800605
#> [37,] -0.32633226 -2.283312279  0.9569437  0.05038291  1.14225380 -2.30443785
#> [38,]  1.68946459  1.713597997 -0.6169675 -0.23887191 -0.37805479  0.03984386
#> [39,]  0.69684199 -0.587290779 -1.0806599 -0.04893910 -1.81322604  0.02863340
#> [40,]  0.50431211 -2.408296181  1.1266273 -1.68601512  1.63554309 -1.11269373
#> [41,]  1.33972255 -1.080841357  0.9567162 -0.04040834 -0.08757263  0.65068265
#> [42,] -1.94570462  0.669407607  0.9232580  0.63069483 -0.65346579 -0.13597110
#> [43,]  1.44064527  0.880437028 -0.5687700  1.23467055 -0.39706674 -0.88948534
#> [44,]  2.90212101  1.750250788 -1.0305339  1.34728294  0.03967435 -0.03530228
#> [45,] -0.91465346  1.098857696  0.4902522  0.11231686  0.05381678  0.23640247
#> [46,]  0.18862750  1.205692066  0.8748675  1.95428319 -0.04751048 -0.61419631
#> [47,] -0.35886467 -0.606824893  0.1892159  1.27444523 -0.76407567 -0.86410225
#> [48,] -0.35654384 -0.705832402 -1.3367786  1.76330989 -0.47787725  0.83398340
#> [49,]  0.86849112  0.270897413  1.5549070  1.66160903  0.78507908  0.38687798
#> [50,] -1.53459518 -0.630037898  2.2925030  2.61898184  1.26419160  0.09823967
#> [51,] -3.41626883  0.603845749  0.6286157 -0.64591993  0.62815707 -0.13630060
#> [52,]  1.33921123 -0.645621319 -0.3529361 -1.14041519  0.93141610 -0.32091502
#> [53,] -2.37952719  0.908613247  1.0620676  0.76765074 -0.22093819 -0.30568717
#> [54,] -1.38198152  1.061000220 -0.5827538 -0.60326340 -2.85129055 -0.92831880
#> [55,] -0.02112469 -0.765740331  1.0213768  0.48006122  0.68121177  0.42094173
#> [56,] -0.14017729  0.403028320  2.5433807 -1.68785846 -1.88256940 -1.36242109
#> [57,] -1.49609868  0.417133595  1.7872431 -0.78494746 -0.45417647  0.37033958
#> [58,]  0.10298247  0.022670091 -1.9887771 -0.49580721  0.92223187  2.37757536
#> [59,]  1.13895881 -2.228722390  0.7411863 -1.57117365 -0.52254147  1.01890428
#> [60,]  0.41308705  0.966502912  0.7935402 -0.54035530  1.83738372  1.57902462
#>               PC7          PC8          PC9         PC10
#>  [1,]  0.98920906 -1.331247680  0.508643904 -0.558712412
#>  [2,] -0.85576561 -1.129282797  0.144336880 -0.768719879
#>  [3,] -0.92551597 -0.373169048  0.186912325  1.097958380
#>  [4,] -1.40512163  1.256106203  0.123238416 -0.350130607
#>  [5,] -0.49146272 -1.344626735 -0.541325297  0.452166422
#>  [6,]  0.72068968 -0.926956564  0.857095527  1.302577771
#>  [7,]  1.43428010  0.479939666  0.973674916 -0.043150399
#>  [8,]  0.90370657  0.798994402 -0.298070142 -0.209618373
#>  [9,] -0.81731662  0.799031333 -0.657299190 -0.254716551
#> [10,] -0.59032850 -0.089061087 -1.105492741 -0.998304398
#> [11,]  0.34511587 -1.287230786 -0.670604158 -0.228742596
#> [12,]  0.07331302  0.139958468 -1.664122606 -0.395238834
#> [13,]  0.39250461 -0.514100049  0.752311126 -0.163175896
#> [14,]  0.24849963  0.311767068  0.969243311 -0.798920908
#> [15,]  1.82835634 -0.124406574 -0.288850652  1.240149949
#> [16,]  0.91412225  0.734282572 -0.008845660  1.182477563
#> [17,] -0.69243206 -0.785266633 -0.356551158 -0.139739180
#> [18,]  1.36635932 -0.365949043  0.840084718  0.589267603
#> [19,]  2.19444004 -0.066299476 -0.451920021 -0.474378407
#> [20,] -0.37835670  1.087099578 -1.308957126  0.256211611
#> [21,]  0.12778261 -1.342554871 -0.849878848  0.330479304
#> [22,]  1.62611592 -0.064568606  0.032445586 -0.645426881
#> [23,] -0.06286973 -0.019728139  0.866785472  0.328579989
#> [24,] -0.96794785  1.390022018  0.293868802 -0.027166406
#> [25,]  0.23894069  0.649384379 -0.872425219  1.008489631
#> [26,] -0.82494359  0.252357125 -0.096754265  0.393249366
#> [27,] -1.16517555  1.749731756  0.138662584 -0.132709829
#> [28,] -0.22586621  0.046764454  0.489808919  0.293159414
#> [29,]  0.87435787  0.716903874 -1.138288982 -0.779501752
#> [30,]  0.89021883 -0.910750827 -0.192355016 -0.203391872
#> [31,] -1.82342798 -0.358241439  0.117821311 -0.001922600
#> [32,] -1.23932477 -2.099763624 -1.242235863 -0.054210012
#> [33,] -1.12139172  2.272347216 -0.155082399 -0.224653294
#> [34,]  0.50583047 -0.239282416  0.492136275 -0.299121892
#> [35,] -1.21780551  0.317335353  0.885792839  1.176967370
#> [36,]  0.23734362  0.136867847 -0.179574396  0.051067803
#> [37,]  0.01639921 -1.405271295  0.005555418 -0.460807981
#> [38,] -0.20249013  0.316473534  0.421572407 -0.381565413
#> [39,] -1.18134303 -1.499746165  0.293351467  1.204602485
#> [40,] -0.22582890  0.182050821  1.889823263 -0.598917736
#> [41,]  0.48130957  0.110606259 -0.401420536 -0.273412410
#> [42,] -0.61589785 -0.611265142  0.379344511 -1.705436185
#> [43,] -0.06432790 -0.005455767  1.584044212 -0.662372354
#> [44,] -0.34898831 -1.107076642 -1.158139239  0.051831807
#> [45,]  1.06804119  0.535589515 -0.005886766  0.346033397
#> [46,] -0.22173727  0.372618645  0.591879902  0.632369191
#> [47,] -0.09313797  0.371046253 -2.057585722  0.233755510
#> [48,] -0.53368590  0.154201217  1.414596275  0.525846445
#> [49,] -0.88012229  0.497272187 -0.012382038  0.139239429
#> [50,]  0.08393584  1.171553486 -1.084392251  0.349883273
#> [51,]  0.57140487  0.437874072 -0.317405141  0.239117305
#> [52,] -0.03010354 -0.443525844  0.228758730 -0.292823145
#> [53,] -0.99623699 -1.765952513  0.641996474 -0.001615218
#> [54,]  1.28129289  0.559939493  0.793314943 -0.595600189
#> [55,]  0.82339140  0.775359245 -0.012639176  0.220031782
#> [56,]  1.01178796  0.111808522 -0.960244171  0.089287760
#> [57,] -0.21173219  0.634829907 -0.364366320 -0.311778138
#> [58,]  0.56317308 -0.211843540  0.506915950 -0.363862621
#> [59,] -0.28074920  0.854726448  1.077541599  0.086772933
#> [60,] -1.12048832  0.197780386 -0.048462965 -0.421729125
#> 
#> $pca_center
#>  [1] -0.19528802  0.35736637  0.20852324  0.02817919  0.08022248 -0.07598368
#>  [7]  0.01316175  0.08751549 -0.05739960 -0.05836311
#> 
#> $pca_scale
#> [1] FALSE
#>